pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apkt-agent"
version = "0.1.0"
description = "APKT data extraction agent"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Development Team"}
]

dependencies = [
    "PyYAML>=6.0",
    "pydantic>=2.0",
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.10.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
apkt-agent = "apkt_agent.cli:main"

[tool.setuptools]
packages = ["apkt_agent"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
apkt_agent = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
